docs: update cloudflare example
